+++ src/luxon.ts
@@ -3,9 +3,9 @@ import Duration from "./duration.js";
 import Interval from "./interval.js";
 import Info from "./info.js";
 import Zone from "./zone.ts";
-import FixedOffsetZone from "./zones/fixedOffsetZone.js";
-import IANAZone from "./zones/IANAZone.js";
-import SystemZone from "./zones/systemZone.js";
+import FixedOffsetZone from "./zones/fixedOffsetZone.ts";
+import IANAZone from "./zones/IANAZone.ts";
+import SystemZone from "./zones/systemZone.ts";
 import Settings from "./settings.js";
 
 const VERSION = "3.7.2";