+++ test.mjs
@@ -1,3 +1,1 @@
 import { Zone } from "./src/luxon.ts";
-
-console.log(Zone());